Clean up the types in the use-profile hook
This involves simplifying some of the cute-siness in useProfile,
specifically, now using `enabled` to run the fetch only if the userId
is present in the useAuth hook's return value. This should always be
true anyway whenever there is a valid session with which to fetch the
profile. But previously the queryFn would run and just return null;
now it will be in a pending state when the user is logged out. But this
should be fine because they shouldn't be able to access these pages!

Everything should be fine, but because the previous approach was there
to paper over some potential edge and corner cases (mismatches between
the actual validity of the cookie and the useAuth hook's sense of it),
it is difficult to test. In any event, the expected behavior should be
more clear now and the edges should be a bit cleaner.