Create client logic to grab scheduled cards + new
Still not quite done yet; this is happening on the review/go page,
but it needs to happen on the previous page when the "let's go" button
gets clicked, and then persisted to localstorage right there and then
the logic here should mostly work. It's set up to be just a little
bit djanky right now, ordering unreviewed cards by "created_at" desc
and picking the first 15. There must be a better way than this, but
we are baby-stepping this one over the finish line.