Use localStorage for persistent storage of today's review
With this update to the /review screens, we are now taking the user's
approval of the day's review, and firing a mutation that stashes the
day's ordered list of Phrase IDs (pids), and a useQuery that fetches
it, using localStorage instead of the database. It feels a little
bit messy still; perhaps we should be using persistQueryClient (but
it's experimental, and based on my very limited knowledge rn, it
wants to stash the entire queryClient, and I'm not sure I'm up for
that right now. The thing works. It is an enhancement over what we
had before it, so we're committing. But we still need to handle:

1. When you navigate away from the page and back again, it should
   know to pick up where you left off. We just have this ordered list
   of pids to study today.
2. For perhaps the same reason, we can't go back and see a previous
   review (let alone edit it).