Lock down the user_deck_plus view
It was initially left open as a way to augment public profiles, which we
have not yet implemented. But it does contain some activity info so we
should lock it down and when we go to implement these more advanced
features for friends/connections, we can make another view using the
prefix `public` as we do for profiles, e.g. `public_profile` and
`public_deck`.