Check if link.Icon is truthy, rather than is function
This is what just recently gave us trouble with the Callout also,
why we had to make everything into a render function. It wasn't because
of the rules of react rendering, it was our boolean check! We've fixed
the erroneous check but have not gone back to change anything about the
Callback, whose _types_ still say they expect a function returning an
IconType or ReactNode, but which actually at runtime should work even
if you just pass a component.

Fixes #95