Show the navbar on scroll up
This one is tricky to get right, compared with just using css
'sticky'. We want to keep the intersection observer but we have
to stop using sticky and instead use position: absolute, and we
need to make it position static whenever we're not scrolling up.

This seems to be working, after one commit. But the transitions
are jarring at the moment and I'm not sure about setting height
or relative position, or anything like that, without a little
more work going into it.