Don't print PhraseCards when no translations
This is actually a hotfix to make up for having bad data in the system;
there should always be a translation for every phrase, or else the phrases
without translations in the correct (for the user) language(s) should be
filtered out either in the query or by the UI/shaping logic -- perhaps
this is a job for the useProcessedPids type approach, as in 088e598,
where we would go through and filter _public info_ from an unauthenticated
query, using private data from a user-authenticated query (user's language
preferences / spoken languages). But for now we'll just make sure not to
hit a runtime error for phrases with no translations.