Add useProcessPids, a useMemo hook for the processPids fn
Note: this is not useProcessedPids, the earlier idea, which would
actually fetch things / ensure the query cache had the data and then
use it. This is useProcessPids, which has the same fn signature as
processPids, but wraps that function in a useMemo only. It's a simpler
approach that doesn't really change the way data is managed; we were
already doing a useMemo each time we called processPids, so we're
just DRY'ing the useMemo as well as the rest of the logic.