Improve styling for the recommended phrases section
This might be even better if we allow the things to stretch a little,
either using flex basis or just checking the length and deciding when
to take up both columns.

Also might be better to cap the height; some things will have more
translations than we care to show. We can just taper off here.