WIP: working up till the mutation
Okay so, here we have redone the processPids to return a bunch of
Sets (maybe I will regret this) so we can easily do set union and
exclusions to make sure the various stages of picking cards do not
interfere with user expectations -- i.e. if a user is given the
chance to select or deselect phrases to add, and they deselect one,
it should probably not get added back in without their permission
on a subsequent step. It seems to be working up till the point of
getting to the Algo recs, but we still need to test the form/mutation
which will actually create the cards, and then give some precedence
to the user-selected cards rather than (as is the current logic)
filling the quota from the deck and then adding the algo-selected,
user-confirmed cards on top of this. But for now it is progress.