WIP: Remove translations we won't understand
This logic is currently living in entirely the wrong place, at the
very last step before rendering one specific set of phrase translations
in the preview for cards to be added from recs. But it is a working
filter, which removes translations into languages we won't understand,
and also removes phrases with zero remaining translations after that
filtering step is complete.

To complete: we will want to move this to much earlier in the logic,
on this page, and possibly into its own hook, akin to how the
`useDeckPidsAndRecs` hook combines data from both useLanguage and
useDeck. We'll want something like perhaps some

```
const { myPhrasesMap, myPids } = useLanguageFilteredForMe(lang)
```

would make sense, to collate a new PhrasesMap type object and the
pids that remain after filtration.