Add localStorage helpers for these "extra" reviews needed
This is a bit preliminary because the review interface already needs
a number of refactors, but at least for now I'm going to assume we
keep using localStorage to track the on-device state of the review
(its manifest and current reviewed-/finished-status), and simple
state for the UI state (which card we're looking at). We may copy
this Route for the "extras" part and change the UI a little bit,
e.g. not having (1 of 20 cards) because this might be annoying when
you're repeating the same 6 cards over and over till you get them.