Make it find 'Again' cards as well as unreviewed ones
This was remarkably simple to get working like this. But now, when
you look at an 'Again' card, it's showing the answer buttons but
not showing the translations. Which is half correct and half incorrect
(should show neither!) and makes me realise that actually when you
click back normally, you get the same behaviour because the buttons
are shown based on _either_ showTranslations _or_ the presence of
mutationData.

So we will have to fix this somehow. These two show/hides should use
the same state or derived state, and it should somehow be reset for
'Again' cards when you get to the end of the list and choose to go
through again.