Combine all the map'd content into one component
This also un-does the sort of wonky thing about having many many
cards mounted and rendered but only one variable for whether to show
the translations and buttons or not. It always seemed like it'd make
more sense to have each card know whether it should be shown or not,
and now that's the way it works. (But that logic hasn't quite been
adjusted yet so this is WIP with a regression.)

And tweaks the <Card> markup in the WhenComplete component so we have
one full self-contained card in that file and don't even have to import
Card into the larger flash-card-review-session, so it's a component
that just handles navigation and control, and each review-single-card
has its own state, mutation, etc.