Create this /again route which has a 'loop' nextCard approach
This is a pretty big commit that duplicates the /go url to /again
and takes flash-card-review-session into a direction where it is
supporting two different approaches in one component. Component
composition is not exactly leveraged to the max here; there are a
bunch of ifs/thens. And this is a WIP but it is nudging the pieces
into place.