Use a useQuery and mutation to manage localStorage
This PR also makes a specific "reviewStage" sort of variable (that
lives inside this useQuery, called useReviewStage, that more explicitly
controls whether we're doing our first review, or second, or we
skipped the second and went to the next interstitial, etc. See type
`ReviewStages` for a description of all 6 values.

We've also made some decisions about how to handle the navigation
between single cards for review. Now all the navigation checks the
original manifest in localStorage; there are no secondary lists being
looped over. And we introduce another navigation option called 'loop'
which will look for the next again card, and if it doesn't find one,
will start from the beginning and look at the rest of the items.

This commit is a kind of minimal working enhancement, but there are
a few obvious extensions to make from here, in subsequent commits:

1. Use a different navigation in "skipped" and "again" mode that just
   lets you double-skip the card (you won't see it again today because
   this mode does not loop)
2. Fix component composition on the different card views being shown
3. Potentially create another route for the WhenComplete pages (there
   are now 4)

Also note: there is one known bug, which is that the query cache keep-
ing track of how many skipped/again cards there are only re-updates
on account of the mutation (pressing a button!) which means that the
first time you see the WhenComplete page is thinks you have X skipped
cards, where X is the number of unreviewed cards back when the
component first mounted (the length of the entire manifest).