Tidy copy and update card counts when the isComplete is true
Note: to make this work, we did have to use a useEffect in a way
that I'm not super happy with. A better approach might be to move
the when-review-complete component to its own route and re-run this
refetch-state-from-local-storage effect in the loader for that page.

Then we would have an approach where the 3 review modes would all
happen on /review/go, and the 3 different completion screens would
all appear on /review/go/complete (or review/complete, makes more
sense as a URL but doesn't slot as easily into the current component
tree structure).

For some time I considered making all of these have their own URLs,
such as /review/go/1-2-3 and /review/complete/1-2-3, but you actually
aren't supposed to be able to access these things through navigation
so much as through the state of the application, so I think it still
makes sense to treat this review screen as a state-based SPA and not
give users much "navigation control". They get buttons like forward
and back, or in stage 2 and 3 they get buttons like "next" and "bury"
and the app manages the rest.