Tidy component structure and add "skip for today" button (stage 2+)
Just to note: we because we're using local state for things, and not
using useQuery hooks for all the things that come from localStorage,
we really have to keep track of when it's okay to unmount and remount
components vs. when we just have to hide them. Hopefully this will be
easier with a second round of use-query-fying the data (moving the
local prevData state / the local ReviewRow into a query and mutation
like we did with the review-state cache, so we can invalidate the
query when we change what's in the cache, and consumers of the hook
will be notified to re-render, rather than simply changing localStorage
without reactivity.

Hopefully a switch to a useStore approach or tanstack/db will also help
us keep this all in one place. Let's see! For now it seems to be mostly
working*.

Todo: There is some bug in the next-card-chooser (perhaps the 'loop'
function) where we're told we have 3 or 4 "again" cards but then
are only presented 2 before being told we're done.