Use react-query AND localStorage for everything
Finally everything seems to work. We are essentially using localStorage
for optimistic updates and then never invalidating them/replacing them
with re-fetched data from the server. We just only set the optimistic
update during the onSuccess function of the mutation.

It's pretty clear after writing this and rewriting it a few times that
we have been muddling throuh two different (as-yet) unfamiliar problems
at the same time: 1) optimistic updates, 2) a store-like state-machine
kind of behavior. But at least the logic is working now; things that
need to be reactive are reactive, everything that updates gets updated
in the DB (if necessary) and in the query cache and in the localStorage.

It seems like now that we're here, a lot of what we're doing that hasn't
been made reactive yet is actually just keeping all the ReviewRows in
localStorage; well we could actually move that all to the database and
use react-query like a normal person. And a lot of the localStorage state
(like the 'review-state' key) is very useStore-like. And it seems clear
that the dailyCacheKey should probably be a context.

But for now, this is all working.

NOTE: We _may_ be experiencing some weird performance issues with the
loading taking a few seconds (it takes much less time when the review is
complete, despite the same async API calls being required).