Replace deck id with lang and card id with phrase id
This is a big and dangerous commit. It includes migrations, changes to
foreign keys, table constraints, seed data, hooks, fetchers, types,
routes and components. But this is the meat of the entire PR and was
not possible to break it up into multiple (working) commits, so it
felt cleaner like this. Nevertheless. This must be tested every which-
a-way.