Add uid explicit UID constraints to views, functions, queries
This is necessary for a few reasons: firstly the user_deck_plus and
user_card_plus were giving wrong results when logged in as super-
user because the aggregates would run and calculate, for example,
all reviews ever done by any user against the phrases in that lang,
rather than just the reviews done by the user who owns the deck.
This wouldn't cause user-facing problems, probably, but it is wrong.

Second, I wanted to be very clear in the RPC functions that we'd
only be selecting and modifying the user's own rows even if RPC
functions handle RLS in ways that might be confusing or might change
down the road.

Thirdly, there was at least one run-time bug with fetchDeck not
working on a very specific deck unless I added an explicit
`.eq('uid', uid)` to the supabase query chain. It's entirely unclear
why that was needed, but I was already researching #138 and just
noticed that adding the (should be redundant) check for uid made
the page work as expected. Explicit .eq checks have been added
in a few other places but I haven't gone through the whole thing.