WIP: Replace ReviewsLoaded w ReviewsMap + another hook for stats
It's not really working still. The review stats hook is working but
there is something wonky either with the gotoNextValid function (having
to use an up-to-date value for the reviewsMap despite the goto fn
being called before it has a chance to refresh), or with my testing
setup: specifically, that when I clear the localStorage to test more,
the Zustand store resets but the server state persists and we're not
using a custom persister (yet) to manage this. Doing so will also
suggest that on the review.index page we should check for an existing
review and use that instead.

This seems like it might be more practical to do that next step in
the process, despite not having planned to do this when we got started
with the feature... or I will solve my testing env issues and defer
that for a later PR.