Add minimally-working Db persistence for the review manifest only
This commit adds a "manual" form of database persistence for the review
manifest only. Originally we were planning to persist the entire store,
but updating every time we update the card index seemed super excessive,
and would immediately constitute some 95% of requests for most review
sessions.

And if we're going to ditch the currentCardIndex, imagine: you start
your review on one device, work on it for about 15 minutes, then you
break for lunch and go to the park with your brownbag. Then you are
munching away and you pull out your phone and load up the session and
now you have to click "go to 'agains'" once to get back to your earlier
place? This is not a big deal. So if we're ditching this very local
form of state, the currentCardIndex, we can ditch the 'stage' too.

`manifestLength` and `dailyCacheKey` are both derived properties.

`reviewsMap` is already being loaded by the page, although it will get
out of sync across devices until we write a little more code in a sub-
sequent commit. So all that was left was the manifest; the only thing
that truly needed to be persisted.

Further... the Zustand persist middleware setup requires functions to
set and get specific data keys (like `stage` and `reviewRows`) but does
not provide a way to set and get multiple keys at once, so when the
`reviewRows` get another row added, and the `currentCardIndex` increments,
we will need two requests just for this. When the page loads it will
presumably attempt to load all fields at once, requiring some 8 different
blocking requests before the page activates. Plus all the other issues
mentioned in the Zustand docs for asynchronous persisters and async
hydration issues.

So since the manifest is the only thing strictly needed here... and it only
needs to be set _once_ per day, it seemed like a much better approach to
just do it like this and actually let the Zustand middleware keep using
localStorage for persistence across browser/app closings/openings.