Update accent and accent-foreground colours
With this change, we are starting to use colour codes in a way that I
think will work more generally. The LangBadge is a badge with a new
variant: 'lang', which uses a `bg-accent text-accent-foreground` in a
way that I think really works! Accent is a soft pastel background colour
and accent-foreground is either dark or light, but still identifiably
green, and looks good against the background variant. Only the badge
variant implements this just as it should be, but the newish button
variant outline-primary and the prior work on destructive-outline are
all attempting to do this same thing nicely, and so the accent and
accent-foreground colours really meshing well together, and having this
badge variant implmentation sitting there ready to copy provides a
potential way forward for the rest of the system.