Improve composite usePhrase + remove filtered phrasesMap
This change removes the phrasesMapFiltered, which was a pretty
big duplication of memory for the client. Instead, we can use
languagePidsFiltered to get a filtered list, and pass pids to
child components when necessary, which can then use the usePhrase
hook to access composite data. We also move two card components
into the `@/components/cards` folder so we can keep shaping these
up together as more of a unit.