Finish getting rid of useDeckPidsAndRecs
This commit completes the work of the last one by removing the recs
from the monolith useDeckPidsAndRecs, and transitioning all consumers
over to using useRecommendations. We then find it's super simple to
stop using most of the other fields because they are copied directly
from deck.pids, so we transition all of those away too. Then there
are only a few derived fields left, like not_in_deck and such, which
require both language data and deck data. Thankfully, these are already
being calculated for useRecommendations, so we include them there
as well and rename the hook to useCompositePids.