Overhaul badges
This commit overhauls our use of badges, removing almost all
className overloads (leaving only a couple that seem appropriate).
Along the way we've greatly simplified the badge display on the
DeckCards that show up on the learn/index page. It's much easier
to look at now, without the different colours and emphasis levels.