Make the Contributions page real
This change renames "requests" (as in, My Requests It  added) to
"contributions" and gives the page tabs for phrases and requests
we authored. It also settles on some new icons for requests and
phrases, to make them sort of co-equal citizens in the language
deck context, and re-orders some of the navs and menus to embody
the spirit of these and coming changes.