Add request upvotes
This commit also removes all mention of meta_request_info because it
wasn't being used for anything anymore, now that we are denormalising
the upvote counts and incremeneting/decrementing the counter directly
on the record like we do with comment upvote count.

And we removed the fulfilled_at because it is either behind the times
or too far ahead of them!